Implemented component interfaces
- Check CHANGELOG.md for more detailed information
